// src/app/(post-login)/submission/program/[shortName]/dashboard/DonorDataSummary/types.tsx
export type CompleteIncompleteFilterCounts = {
	completed: number;
	incomplete: number;
	noData: number;
};

export type ValidInvalidFilterCounts = {
	valid: number;
	invalid: number;
};

export type TumorNormalStatusCounts = {
	tumorAndNormal: number;
	tumorOrNormal: number;
	noData: number;
};

export type TumorNormalMatchedPairStatusCounts = {
	tumorNormalMatchedPair: number;
	tumorNormalNoMatchedPair: number;
	tumorNormalMatchedPairMissingRawReads: number;
	noData: number;
};

export type CompletedInProgressFailedFilterCounts = {
	completed: number;
	inProgress: number;
	failed: number;
	noData: number;
};

export type DataSubmittedDataNotSubmittedFilterCounts = {
	dataSubmitted: number;
	noDataSubmitted: number;
};